Make comparison of atom and property names case sensitive.
MP4Track::FinishSdtp looked for the ftyp atom, but due to the case insensitive comparison found an atom named Ftyp (which is treated as an unknown atom). It then casted this unknown atom to type MP4FtypAtom and tried to access its properties, which are not present in unknown atoms.

Addresses: https://nvd.nist.gov/vuln/detail/CVE-2018-17235